Rename "super" key to "meta"
This is inconsistent with the W3C spec, and while it's arguably not the
best name, it's worse that Winit is diverging and choosing a different
name.

---

Full list of changes:

KeyCode:
- SuperLeft -> MetaLeft
- SuperRight -> MetaRight
- Meta -> Super

NamedKey::Meta is swapped with NamedKey::Super.

ModifiersState::SUPER is renamed to ModifiersState::META.

ModifiersState::super_key is renamed to ModifiersState::meta_key.

ModifiersKeys::LSUPER and RSUPER are renamed to LMETA and RMETA.